+++ src/components/TodoInfo/TodoInfo.tsx
@@ -1,4 +1,12 @@
 // Add the required types and props
-export const TodoInfo = () => (
-  <>TodoInfo markup</>
+import React from 'react';
+
+import { Todo } from '../../types/Todo';
+
+interface Props {
+  todos: Todo;
+}
+
+export const TodoInfo: React.FC<Props> = ({ todos }) => (
+  <h2 className="TodoInfo__title">{todos.title}</h2>
 );
